forms/translation.py
from modeltranslation.translator import register, TranslationOptions
from forms.models.model_contact_us import ContactUs


@register(ContactUs)
class ContactUsTranslationOptions(TranslationOptions):
    fields = ('name', 'message', 'is_active')
