requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.5
django-ckeditor==6.5.1
django-environ==0.9.0
django-filter==22.1
django-jalali==6.0.1
django-js-asset==2.0.0
django-modeltranslation==0.18.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
inflection==0.5.1
jdatetime==4.1.0
jsonschema==4.17.3
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
